refactor(react): removed further react abstractions [PW-355]
