--- frontend/admin/src/components/BackButton.jsx
@@ -1,11 +1,0 @@
-import "../css/back-button.css";
-
-import FaArrowLeftLong from "./Icons/FaArrowLeftLong";
-
-export default function BackButton({ href, text }) {
-    return (
-        <a href={href} className="hoverable back-button-anchor">
-            <FaArrowLeftLong /> Back to {text}
-        </a>
-    );
-}
