+++ frontend/admin/src/components/UnauthorizedRoute.jsx
@@ -1,19 +1,15 @@
 import { useEffect, useState } from "react";
-import { useNavigate } from "react-router-dom";
-
 import Loading from "../pages/loading";
-
-import { validateToken } from "../services/auth";
+import { validateToken } from "../js/auth-service";
 
 export default function UnauthorizedRoute({ children }) {
-    const navigate = useNavigate();
     const [loading, setLoading] = useState(true);
 
     useEffect(() => {
         (async () => {
             try {
                 await validateToken();
-                navigate("/blog");
+                window.location.href = "/blog";
             } catch {
                 localStorage.removeItem("token");
                 setLoading(false);
