--- frontend/admin/src/css/blog-header.css
@@ -1,17 +1,0 @@
-.header-wrapper {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-}
-
-.header-logout-btn {
-    font-size: 12px;
-    color: rgb(212 212 212);
-}
-
-.header-logout-btn-content {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-    font-size: 12px;
-}
