+++ frontend/admin/src/css/blog.css
@@ -16,3 +16,21 @@
     margin-bottom: 6px;
     font-weight: 500;
 }
+
+header {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
+
+.header-logout-btn {
+    font-size: 12px;
+    color: rgb(212 212 212);
+}
+
+.header-logout-btn-content {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+    font-size: 12px;
+}
