--- frontend/admin/src/css/footer.css
@@ -1,20 +1,0 @@
-.footer-wrapper {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 12px;
-    font-size: 10px;
-}
-
-.footer-absolute {
-    position: absolute;
-    left: 0px;
-    right: 0px;
-    bottom: 16px;
-}
-
-.footer-relative {
-    margin-top: 64px;
-    padding-bottom: 16px;
-}
