+++ frontend/admin/src/js/auth-util.js
@@ -1,0 +1,4 @@
+export default function logout() {
+    localStorage.removeItem("token");
+    window.location.href = "/";
+}
