+++ frontend/admin/src/pages/not-found.jsx
@@ -1,12 +1,15 @@
+import Footer from "../components/Footer";
 import "../css/not-found.css";
-import Container from "../components/Container";
 
 export default function NotFound() {
     return (
-        <Container absoluteFooter>
-            <section className="center">
-                <p className="not-found-text">Not Found</p>
-            </section>
-        </Container>
+        <main>
+            <div className="content">
+                <section className="center">
+                    <p className="not-found-text">Not Found</p>
+                </section>
+            </div>
+            <Footer />
+        </main>
     );
 }
