+++ frontend/main/src/components/Footer.jsx
@@ -1,9 +1,9 @@
 export default function Footer() {
     return (
         <footer>
-            <section className="footer-links">
+            <div className="footer-links">
                 <a
-                    aria-label="Github"
+                    aria-label="github"
                     href="https://github.com/michaelhyi/"
                     rel="noopener noreferrer"
                     target="_blank"
@@ -39,7 +39,7 @@ export default function Footer() {
                         <path d="M880 112H144c-17.7 0-32 14.3-32 32v736c0 17.7 14.3 32 32 32h736c17.7 0 32-14.3 32-32V144c0-17.7-14.3-32-32-32zM349.3 793.7H230.6V411.9h118.7v381.8zm-59.3-434a68.8 68.8 0 1 1 68.8-68.8c-.1 38-30.9 68.8-68.8 68.8zm503.7 434H675.1V608c0-44.3-.8-101.2-61.7-101.2-61.7 0-71.2 48.2-71.2 98v188.9H423.7V411.9h113.8v52.2h1.6c15.8-30 54.5-61.7 112.3-61.7 120.2 0 142.3 79.1 142.3 181.9v209.4z" />
                     </svg>
                 </a>
-            </section>
+            </div>
             <p>&copy; 2023 Michael Yi, All Rights Reserved.</p>
         </footer>
     );
