+++ frontend/main/src/css/portfolio.css
@@ -108,5 +108,3 @@
     padding-left: 4px;
     font-size: 10px;
 }
-
-
