+++ frontend/main/src/index.css
@@ -17,7 +17,7 @@
 }
 
 main {
-    color: rgb(255 255 255);
+    color: white;
     background-color: rgb(23 23 23);
     display: flex;
     flex-direction: column;
